HSEARCH-3115 Move classes not strictly implementing the DSL out of the DSL package
Let's keep only DSL interfaces in there, and let's not mix them with
classes that are supposed to make use of this DSL, or be used by the
DSL.
