+++ backend/elasticsearch/src/main/java/org/hibernate/search/backend/elasticsearch/analysis/ElasticsearchAnalysisConfigurer.java
@@ -4,8 +4,9 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.search.backend.elasticsearch.analysis.model.dsl;
+package org.hibernate.search.backend.elasticsearch.analysis;
 
+import org.hibernate.search.backend.elasticsearch.analysis.model.dsl.ElasticsearchAnalysisDefinitionContainerContext;
 import org.hibernate.search.backend.elasticsearch.cfg.SearchBackendElasticsearchSettings;
 
 /**
