+++ backend/lucene/src/main/java/org/hibernate/search/backend/lucene/analysis/LuceneAnalysisConfigurer.java
@@ -4,8 +4,9 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.search.backend.lucene.analysis.model.dsl;
+package org.hibernate.search.backend.lucene.analysis;
 
+import org.hibernate.search.backend.lucene.analysis.model.dsl.LuceneAnalysisDefinitionContainerContext;
 import org.hibernate.search.backend.lucene.cfg.SearchBackendLuceneSettings;
 
 /**
