+++ backend/lucene/src/main/java/org/hibernate/search/backend/lucene/analysis/model/impl/LuceneAnalysisDefinitionRegistry.java
@@ -4,7 +4,7 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.search.backend.lucene.analysis.model.dsl.impl;
+package org.hibernate.search.backend.lucene.analysis.model.impl;
 
 import org.hibernate.search.backend.lucene.analysis.model.dsl.impl.annotations.AnalyzerDef;
 import org.hibernate.search.backend.lucene.analysis.model.dsl.impl.annotations.NormalizerDef;
