+++ backend/lucene/src/main/java/org/hibernate/search/backend/lucene/analysis/model/impl/SimpleLuceneAnalysisDefinitionRegistry.java
@@ -4,7 +4,7 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.search.backend.lucene.analysis.model.dsl.impl;
+package org.hibernate.search.backend.lucene.analysis.model.impl;
 
 import java.lang.invoke.MethodHandles;
 import java.util.Collections;
@@ -13,8 +13,6 @@ import java.util.TreeMap;
 
 import org.hibernate.search.backend.lucene.analysis.model.dsl.impl.annotations.AnalyzerDef;
 import org.hibernate.search.backend.lucene.analysis.model.dsl.impl.annotations.NormalizerDef;
-import org.hibernate.search.backend.lucene.analysis.model.impl.LuceneAnalysisDefinitionCollector;
-import org.hibernate.search.backend.lucene.analysis.model.impl.LuceneAnalysisDefinitionContributor;
 import org.hibernate.search.backend.lucene.logging.impl.Log;
 import org.hibernate.search.util.impl.common.LoggerFactory;
